Renamed -Options type to -Verb type
