+++ tests/dotnet.podcast.tests/GlobalUsings.cs
@@ -1,5 +1,5 @@
 global using Xunit;
-global using dotnet.podcast.options;
+global using dotnet.podcast.verbs;
 global using Microsoft.Extensions.Logging;
 global using System.IO.Abstractions.TestingHelpers;
 global using dotnet.podcast.handlers;
